điều chỉnh tên biến của các entity
Chuyển hết các giá trị của ID từ Long sang Integer

+++ src/main/java/com/poly/demo/entity/TicketFoodId.java
@@ -12,6 +12,6 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TicketFoodId implements Serializable {
-    private Long ticketId;
-    private Long foodItemId;
+    private Integer ticketId;
+    private Integer foodItemId;
 }
